Move cmds/upspin to cmds/upspin_sos
