+++ cmds/upspin_sos/upspin_sos.go
